feat: add rpc typing with union of functions definitions
