+++ test/index.test.ts
@@ -5,3 +5,4 @@ import './resource-embedding'
 import './transforms'
 import './get_username_and_status_rpc'
 import './embeded_functions_join'
+import './advanced_rpc'
